Completed moving client to separate crate
