+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 members = [
     "imgurs-model",
-#    "imgurs-client"
+    "imgurs-client"
 ]
 exclude = [
     ".github"
