+++ imgurs-client/src/endpoints/authorization.rs
