+++ imgurs-client/src/endpoints/gallery.rs
