+++ imgurs-client/src/lib.rs
@@ -1,2 +1,5 @@
-pub mod utilities;
 pub mod client;
+pub mod error;
+pub mod response;
+pub mod traits;
+pub mod utilities;
