--- src/api/mod.rs
@@ -1,21 +1,0 @@
-//! API wrapper
-
-pub mod traits;
-pub mod error;
-pub mod client;
-pub mod response;
-
-pub mod endpoints {
-    //! API endpoints implementations
-
-
-
-    pub mod authorization;
-    pub mod account;
-    pub mod gallery;
-}
-
-
-
-
-
