--- src/model/mod.rs
@@ -1,19 +1,0 @@
-//! API reference model
-
-pub mod account;
-pub mod account_settings;
-pub mod album;
-pub mod basic;
-pub mod comment;
-pub mod conversation;
-pub mod custom_gallery;
-pub mod gallery_album;
-pub mod gallery_image;
-pub mod gallery_profile;
-
-pub mod image;
-
-pub mod message;
-
-pub mod common;
-pub mod authorization;
